Refactor functions to get word info from dictionary and use in StudyText component
