+++ src/utilities/words-service.js
@@ -1,0 +1,24 @@
+export function getWordInfo(word, savedWords) {
+  let pinyin = ''
+  let meaning = ''
+
+  if (word.matches && word.matches[0]) {
+    pinyin = word.matches[0].pinyinPretty
+    meaning = word.matches[0].english
+    meaning = meaning.includes('/') ? meaning.split('/')[0].trim() : meaning
+
+  } else {
+    pinyin = ''
+    meaning = ''
+  }
+
+  return { pinyin, meaning }
+}
+
+export function checkSpecialChar(word) {
+
+  const specialChars = ['‘', '“', '《', '『', '【', '（', '’', '”', '》', '』', '】', '）', '、', '，', '…', '。', '：', '；', '！', '？', '「', '」', '.', '・']
+    const isSpecialChar = specialChars.includes(word.text) || /\d/.test(word.text) || /[^\u4e00-\u9fa5]/.test(word.text)
+
+  return { isSpecialChar }
+}
